# Reduce number of "stat" syscalls for MergeTree data parts