--- a/dbms/src/Storages/MergeTree/MergeTreeReadPool.cpp
+++ b/dbms/src/Storages/MergeTree/MergeTreeReadPool.cpp
@@ -129,12 +129,12 @@ MergeTreeReadTaskPtr MergeTreeReadPool::getTask(const size_t min_marks_to_read, 
         prewhere_info && prewhere_info->remove_prewhere_column, per_part_should_reorder[part_idx], std::move(curr_task_size_predictor));
 }
 
-MarkRanges MergeTreeReadPool::getRestMarks(const std::string & part_path, const MarkRange & from) const
+MarkRanges MergeTreeReadPool::getRestMarks(const MergeTreeDataPart & part, const MarkRange & from) const
 {
     MarkRanges all_part_ranges;
     for (const auto & part_ranges : parts_ranges)
     {
-        if (part_ranges.data_part->getFullPath() == part_path)
+        if (part_ranges.data_part.get() == &part)
         {
             all_part_ranges = part_ranges.ranges;
             break;
@@ -142,7 +142,7 @@ MarkRanges MergeTreeReadPool::getRestMarks(const std::string & part_path, const 
     }
     if (all_part_ranges.empty())
         throw Exception("Trying to read marks range [" + std::to_string(from.begin) + ", " + std::to_string(from.end) + "] from part '"
-                        + part_path + "' which has no ranges in this query", ErrorCodes::LOGICAL_ERROR);
+            + part.getFullPath() + "' which has no ranges in this query", ErrorCodes::LOGICAL_ERROR);
 
     auto begin = std::lower_bound(all_part_ranges.begin(), all_part_ranges.end(), from, [] (const auto & f, const auto & s) { return f.begin < s.begin; });
     if (begin == all_part_ranges.end())