--- a/dbms/src/Storages/MergeTree/MergeTreeReadPool.h
+++ b/dbms/src/Storages/MergeTree/MergeTreeReadPool.h
@@ -81,7 +81,7 @@ public:
     void profileFeedback(const ReadBufferFromFileBase::ProfileInfo info);
 
     /// This method tells which mark ranges we have to read if we start from @from mark range
-    MarkRanges getRestMarks(const std::string & part_path, const MarkRange & from) const;
+    MarkRanges getRestMarks(const MergeTreeDataPart & part, const MarkRange & from) const;
 
     Block getHeader() const;