--- a/dbms/src/Storages/MergeTree/MergeTreeReader.cpp
+++ b/dbms/src/Storages/MergeTree/MergeTreeReader.cpp
@@ -160,7 +160,7 @@ void MergeTreeReader::addStreams(const String & name, const IDataType & type,
         if (streams.count(stream_name))
             return;
 
-        bool data_file_exists = data_part->checksums.files.count(path + stream_name + DATA_FILE_EXTENSION);
+        bool data_file_exists = data_part->checksums.files.count(stream_name + DATA_FILE_EXTENSION);
 
         /** If data file is missing then we will not try to open it.
           * It is necessary since it allows to add new column to structure of the table without creating new files for old parts.