--- a/dbms/tests/performance/merge_tree_many_partitions.xml
+++ b/dbms/tests/performance/merge_tree_many_partitions.xml
@@ -1,0 +1,30 @@
+<test>
+    <name>merge_tree_many_partitions</name>
+    <type>loop</type>
+
+    <create_query>CREATE TABLE bad_partitions (x UInt64) ENGINE = MergeTree PARTITION BY x ORDER BY x</create_query>
+    <fill_query>INSERT INTO bad_partitions SELECT * FROM numbers(10000)</fill_query>
+
+    <stop_conditions>
+        <all_of>
+            <iterations>5</iterations>
+            <min_time_not_changing_for_ms>10000</min_time_not_changing_for_ms>
+        </all_of>
+        <any_of>
+            <iterations>100</iterations>
+            <total_time_ms>60000</total_time_ms>
+        </any_of>
+    </stop_conditions>
+
+    <main_metric>
+        <min_time/>
+    </main_metric>
+
+    <settings>
+        <max_partitions_per_insert_block>0</max_partitions_per_insert_block>
+    </settings>
+
+    <query>SELECT count() FROM bad_partitions</query>
+
+    <drop_query>DROP TABLE IF EXISTS bad_partitions</drop_query>
+</test>