For changelog. Remove if this is non-significant change.

Category (leave one):
- Performance Improvement

Short description (up to few sentences):
Improved performance of MergeTree tables on very slow filesystems by reducing number of `stat` syscalls.