@blinkov 2019-06-16
@jackpgao could you test this patch to see if it helps with partition deletion issues you were talking about?