@alexey-milovidov 2019-06-16
I don't see much difference on my tests: `stat` syscalls on ext4 are very cheap.

```
CREATE TABLE bad_partitions (a UInt64, b UInt64, c UInt64, d UInt64, e UInt64, f UInt64, g UInt64, h UInt64, i UInt64, j UInt64, k UInt64, l UInt64, m UInt64, n UInt64, o UInt64, p UInt64, q UInt64, r UInt64, s UInt64, t UInt64, u UInt64, v UInt64, w UInt64, x UInt64, y UInt64, z UInt64) ENGINE = MergeTree PARTITION BY x ORDER BY x;
INSERT INTO bad_partitions (x) SELECT * FROM numbers(10000);

SELECT count() FROM bad_partitions;

Was:

QPS: 1.189, RPS: 11893.041, MiB/s: 0.091, result RPS: 1.189, result MiB/s: 0.000.
0.000%  0.591 sec.
10.000% 0.596 sec.
20.000% 0.618 sec.
30.000% 0.630 sec.
40.000% 0.652 sec.
50.000% 0.745 sec.
60.000% 0.786 sec.
70.000% 0.800 sec.
80.000% 1.070 sec.
90.000% 1.243 sec.
95.000% 1.430 sec.
99.000% 1.581 sec.
99.900% 1.615 sec.
99.990% 1.619 sec.

After:

QPS: 1.523, RPS: 15228.830, MiB/s: 0.116, result RPS: 1.523, result MiB/s: 0.000.
0.000%  0.495 sec.
10.000% 0.577 sec.
20.000% 0.616 sec.
30.000% 0.627 sec.
40.000% 0.647 sec.
50.000% 0.653 sec.
60.000% 0.670 sec.
70.000% 0.688 sec.
80.000% 0.707 sec.
90.000% 0.724 sec.
95.000% 0.754 sec.
99.000% 0.772 sec.
99.900% 0.775 sec.
99.990% 0.775 sec.
```