@alexey-milovidov 2019-06-16
But it will definitely help @filimonov to test ClickHouse on S3 over FUSE (that's not a good idea BTW).