@filimonov 2019-06-17
For benchmarks - try `select * from hugetable limit 1` (the visible delay on ext4 with such selects looks pretty the same as almost any select on fuse).