@filimonov 2019-06-17
Same sample table, Ext4, SSD, `select * from bad_partitions limit 1`.

Before:
```
echo 'select * from bad_partitions limit 1' | clickhouse-benchmark -i 20 -d 0
Loaded 1 queries.

Queries executed: 20.

QPS: 0.685, RPS: 0.685, MiB/s: 0.000, result RPS: 0.685, result MiB/s: 0.000.
0.000%	1.427 sec.
10.000%	1.428 sec.
20.000%	1.429 sec.
30.000%	1.431 sec.
40.000%	1.433 sec.
50.000%	1.434 sec.
60.000%	1.436 sec.
70.000%	1.439 sec.
80.000%	1.443 sec.
90.000%	1.467 sec.
95.000%	1.529 sec.
99.000%	1.701 sec.
99.900%	1.740 sec.
99.990%	1.744 sec.
```

After:
```
echo 'select * from bad_partitions limit 1' | clickhouse-benchmark -i 20 -d 0
Loaded 1 queries.

Queries executed: 20.

QPS: 3.109, RPS: 3.109, MiB/s: 0.001, result RPS: 3.109, result MiB/s: 0.001.
0.000%	0.304 sec.
10.000%	0.306 sec.
20.000%	0.307 sec.
30.000%	0.307 sec.
40.000%	0.308 sec.
50.000%	0.309 sec.
60.000%	0.310 sec.
70.000%	0.311 sec.
80.000%	0.314 sec.
90.000%	0.315 sec.
95.000%	0.323 sec.
99.000%	0.424 sec.
99.900%	0.447 sec.
99.990%	0.450 sec.
```

About 4.5 times better in that case.