@filimonov 2019-06-18
On s3 + fuse (goofys 0.20) performance is up to 3 times better than w/o that change. 

Some samples:
```
select count() from tripdata

Before: 1 rows in set. Elapsed: 13.919 sec. Processed 1.31 billion rows, 1.31 GB (93.98 million rows/s., 93.98 MB/s.)
After: 1 rows in set. Elapsed: 6.330 sec. Processed 1.31 billion rows, 1.31 GB (206.66 million rows/s., 206.66 MB/s.) 

SELECT avg(passenger_count) FROM tripdata

Before: 1 rows in set. Elapsed: 14.075 sec. Processed 1.31 billion rows, 1.31 GB (92.93 million rows/s., 92.93 MB/s.)
After: 1 rows in set. Elapsed: 5.261 sec. Processed 1.31 billion rows, 1.31 GB (248.62 million rows/s., 248.62 MB/s.) 

SELECT avg(passenger_count) FROM tripdata  WHERE toYear(pickup_date) = 2016

Before: 1 rows in set. Elapsed: 1.917 sec. Processed 131.17 million rows, 393.50 MB (68.42 million rows/s., 205.27 MB/s.)
After: 1 rows in set. Elapsed: 1.385 sec. Processed 131.17 million rows, 393.50 MB (94.70 million rows/s., 284.11 MB/s.) 
```