@alexey-milovidov 2019-06-18
> Same sample table, Ext4, SSD, select * from bad_partitions limit 1.

> About 4.5 times better in that case.

It because I had also lowered CPU usage. Before, the query was bounded by userspace CPU, not by syscalls.