@filimonov 2019-06-18
> > Same sample table, Ext4, SSD, select * from bad_partitions limit 1.
> 
> > About 4.5 times better in that case.
> 
> It because I had also lowered CPU usage. Before, the query was bounded by userspace CPU, not by syscalls.

Ok. Important thing that it's now faster.

BTW: is there some good reason why it still takes 0.3 sec to read one row? (max_block_size and max_threads doesn't help)