@den-crane 2019-06-18
19.4.3.11. HDD raid 10

INSERT INTO bad_partitions SELECT * FROM numbers(10000)

```
echo 'select * from bad_partitions limit 1' | clickhouse-benchmark --database dw  -i 20 -d 0
Loaded 1 queries.

Queries executed: 20.

QPS: 16.449, RPS: 16.449, MiB/s: 0.000, result RPS: 16.449, result MiB/s: 0.000.
0.000%	0.054 sec.
10.000%	0.054 sec.
20.000%	0.055 sec.
30.000%	0.055 sec.
40.000%	0.055 sec.
50.000%	0.055 sec.
60.000%	0.055 sec.
70.000%	0.055 sec.
80.000%	0.056 sec.
90.000%	0.057 sec.
95.000%	0.061 sec.
99.000%	0.063 sec.
99.900%	0.064 sec.
99.990%	0.064 sec.
```